gs-reactor: ApplicationServer: Avoid null pointer exception
Check that the root DOM node is not null before destroying the associated
context when the websocket is closed. The root DOM node can be null if the page
is reloaded before it was completely built.

+++ gs-reactor/src/main/java/org/genericsystem/reactor/appserver/ApplicationServer.java
@@ -193,7 +193,8 @@ public class ApplicationServer extends AbstractBackEnd {
 			return (v) -> {
 				for (CacheSocket cacheSocketContext : caches)
 					if (socket.equals(cacheSocketContext.getSocket())) {
-						cacheSocketContext.getDomNodeVerticle().getRootHtmlDomNode().getModelContext().destroy();
+						if (cacheSocketContext.getDomNodeVerticle().getRootHtmlDomNode() != null)
+							cacheSocketContext.getDomNodeVerticle().getRootHtmlDomNode().getModelContext().destroy();
 						cacheSocketContext.getCache().disableReactive();
 						caches.remove(cacheSocketContext);
 					}
